requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
assemblyai==0.42.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
elevenlabs==2.9.1
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
murf==2.0.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
